+++ src/dcat-us/index.ts
@@ -1,9 +1,10 @@
+import { IModel } from '@esri/hub-common';
 import { listDependencies } from 'adlib';
 import { buildDatasetTemplate, DcatDatasetTemplate, formatDcatDataset } from './dataset-formatter';
 import { FeedFormatterStream } from './feed-formatter-stream';
 import { DISTRIBUTION_DEPENDENCIES } from './_generate-distributions';
 
-export function getDataStreamDcatUs11(hostname: string, dcatCustomizations?: DcatDatasetTemplate) {
+export function getDataStreamDcatUs11(siteModel: IModel, dcatCustomizations?: DcatDatasetTemplate) {
   const catalogStr = JSON.stringify({
       '@context':
         'https://project-open-data.cio.gov/v1.1/schema/catalog.jsonld',
@@ -22,13 +23,15 @@ export function getDataStreamDcatUs11(hostname: string, dcatCustomizations?: Dca
   const datasetTemplate = buildDatasetTemplate(dcatCustomizations);
 
   const formatFn = (chunk) => {
-    return formatDcatDataset(chunk, hostname, datasetTemplate);
+    return formatDcatDataset(chunk, siteModel, datasetTemplate);
   };
 
   return {
     stream: new FeedFormatterStream(header, footer, ',\n', formatFn),
     dependencies: [
       'id', // used for the dataset landing page URL
+      'type', // used for the dataset landing page URL
+      'slug', // used for the dataset landing page URL
       'licenseInfo', // required for license resolution
       'structuredLicense', // required for license resolution
       ...DISTRIBUTION_DEPENDENCIES,